+++ src/ecs/mod.rs
@@ -1,6 +1,7 @@
 //module tree
 mod auto_despawn;
 mod callbacks;
+mod err;
 mod named_syscall;
 mod spawned_syscall;
 mod syscall;
@@ -8,6 +9,7 @@ mod syscall;
 //API exports
 pub use auto_despawn::*;
 pub use callbacks::*;
+pub use err::*;
 pub use named_syscall::*;
 pub use spawned_syscall::*;
 pub use syscall::*;