+++ src/lib.rs
@@ -6,6 +6,7 @@ use crate as bevy_cobweb;
 //module tree
 pub mod ecs;
 pub mod react;
+pub mod result;
 
 //API exports
 pub use bevy_cobweb_derive::*;
@@ -15,4 +16,5 @@ pub mod prelude
     pub use crate::*;
     pub use crate::ecs::*;
     pub use crate::react::*;
+    pub use crate::result::*;
 }