+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.15.0]
+
+### Changed
+
+- Rename: `ReactorResult` -> `CobwebResult`.
+- Add `CobwebEcsError` and `CobwebReactError` and use them in crate APIs.
+
 ## [0.14.3]
 
 ### Changed